+++ examples/web/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @workglow/web
 
+## 0.3.7
+
 ## 0.3.6
 
 ### Chores